Add intermediate FilterLink container and Link
Having one top-level container and lots of presentational components below it
means that lots of props have to be propagated down the tree, even when
intermediate components don't use them. FilterLink now acts as a container and
subscribes directly to the store, so Footer does not need to pass props, while
Link is a separate presentational component.